Add vim-move plugin
To move blocks/lines up and down.

Ideally, I wanted to map this to A-j etc. Unfortunaly, mapping alt keys
is difficult with vim+roxterm, since roxterm will send keys in such a
way that vim does not know if it was Esc followed by k or Alt+k.
Therefore I mapped A-K etc to not get into a situation where escape
followed by j/k would be interpreted as moving lines.

I concidered switching from roxterm to urxvt, but given that roxterm has
a recent open issue related to this, I'm going to hold off on that for a
while: https://sourceforge.net/p/roxterm/feature-requests/67/